Contributer/requirements.txt
flask==2.3.3
web3==6.11.1
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.32.3
requests==2.32.3